SharedTokenCacheCredential uses AadClient
